functions/requirements.txt
firebase_functions~=0.1.1
python-dotenv~=1.0.0
requests~=2.31.0
google-auth~=2.22.0